refactor: Move UserAdminService of model registration to model security
